requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
line-bot-sdk>=3.0.0
python-dotenv>=1.0.0
google-genai>=0.6.0
